feat: Created variables, mounted hook and setupDeck method
Created the suits array which holds the card suits, the values array
which holds card values, the deck array which starts empty but is
created by looping through the suits and values arrays in the setupDeck
method. Created the mounted hook which runs the setupDeck method when
App.vue is mounted.